+++ client-sdk/ts-web/core/src/registry.ts
@@ -205,7 +205,9 @@ export function registerEntityWrapper() {
 }
 
 export function deregisterEntityWrapper() {
-    return new consensus.TransactionWrapper<void>(METHOD_DEREGISTER_ENTITY);
+    return new consensus.TransactionWrapper<types.RegistryDeregisterEntity>(
+        METHOD_DEREGISTER_ENTITY,
+    );
 }
 
 export function registerNodeWrapper() {