+++ client-sdk/ts-web/core/docs/changelog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased changes:
 
+Breaking changes:
+
+- The `NodeAddress` type is replaced with the equivalent `NetTCPAddr`.
+
 Bug fixes:
 
 - Fixes to type declarations that were inconsistent with the Go types.